Combo Product In POS V16
Created an boolean field in the product with name  'Is Combo' and added a page and the products. Also mention if the products are optional or required, if it is optional then set the count to choose the products from POS. 
If combo products selected from the POS then the products can be view in the orderlines and also in the receipt.